Fixed mis-processing "quiet" commandreturns
When processing multi-command workflows, where some of the commands are
"quiet", the commandreturns were still being processed, even if the
output was mostly suppressed. This resulted in some errant extra commas
in JSON output, for example. 

NOTE: Also discovered is that for multi-command workflows, if using JSON
output, as coded now, the JSON is invalid. Because of an inability to
predict "quiet" commands in the sequence, no commas are being emitted
between document blocks for the last host. e.g. { ... }{ ... } instead
of { ... }, { ... }. This will take some thinking to resolve.