Changes to accomodate Windows builds
Grabbing open files, and obtaining the open file limit are exclusively
UNIXy things, and their inclusion prevented building on Windows. All
that jazz moved to misc_unix.go. misc_win.go is stubbed with the
required function, and returns the GOMAXPROCS value in lieu of
something more awesome if -max/maxexecs is set to 0.